requirements.txt
# Core
django==4.1

# Get environment variables
django-environ==0.9.0

# DB Postgres
psycopg2-binary==2.9.3

#For the admin panel.
django-grappelli==3.0.3

# For contacts generating.
Faker==14.0.0

# For phone numbers generating.
phonenumbers==8.12.53
django-phonenumber-field==6.3.0

# For birthday generating.
django-birthday==0.1.4

# For nickname generating.
nickname_generator==1.0.1

# ...
django-picklefield==3.1

# [code_style]-[BEGIN]
# Pre-commit hook.
pre-commit==2.20.0
# Upgrade syntax.
pyupgrade==2.37.3
# Formatter. Apply codestyle.
black==22.6.0
# Static analysis
prospector[with_everything]==1.7.7
# [code_style]-[END]
